LayoutTests/canvas/philip/tests/2d.drawImage.nonfinite-expected.txt

< [index] >
2d.drawImage.nonfinite
drawImage() with Infinity/NaN is ignored
References: 2d.nonfinite
Actual output:
Expected output:

Aborted with exception: INDEX_SIZE_ERR: DOM Exception 1
